cypress type fix
